+++ portfolio/src/Utils/biografia.js
@@ -1,0 +1,20 @@
+export const biografia = `
+const biografia = [
+  {
+    Título: "Inicio en el Desarrollo Web",
+    Descripción:
+      "Actualmente me desempeño en el área del desarrollo web, tanto Backend como Frontend. Inicié mi carrera en este emocionante campo con un fuerte interés en construir soluciones tecnológicas.",
+    Año: 2022
+  },
+  {
+    Título: "Pasión por el Trabajo en Equipo",
+    Descripción:
+      "Me apasiona el trabajo en equipo, donde todos compartimos un mismo objetivo y nos desempeñamos arduamente en pos de conseguirlo. Creo firmemente en la colaboración y la sinergia para lograr resultados excepcionales.",
+  },
+  {
+    Título: "Explorando Nuevas Tecnologías",
+    Descripción:
+      "El mundo de la programación es increíble; nada me divierte más que seguir aprendiendo tecnologías nuevas e implementarlas con el fin de alcanzar mis metas. Siempre estoy en busca de desafíos y oportunidades para mejorar mis habilidades.",
+  },
+];
+`;
