+++ portfolio/src/index.css
@@ -66,7 +66,7 @@ code {
 
 @media only screen and (max-width: 768px) {
   .custom-carousel img {
-   width: 2200px;
+   width: 3200px;
    height: 120px;
   }
 }
